lib-src/libsndfile: fix warning about format argument type mismatch

+++ lib-src/libsndfile/src/broadcast.c
@@ -181,7 +181,7 @@ gen_coding_history (char * added_history, int added_history_max, const SF_INFO *
 
 	count = snprintf (added_history, added_history_max,
 							"A=PCM,F=%u,W=%hu,M=%s,T=%s-%s\r\n",
-							psfinfo->samplerate, width, chnstr, PACKAGE, VERSION) ;
+							psfinfo->samplerate, (unsigned short)width, chnstr, PACKAGE, VERSION) ;
 
 	if (count >= added_history_max)
 		return 0 ;
